Improve consistency of WeightVector usage
Since all that WeightVector is doing is holding a hvector defaultdict
instance as an attribute and wrapping some of the __methods__, it should
be a subclass of defaultdict. This change also allows it to be used as a
defaultdict instead of other pieces of code accessing its .data_dict
attribute directly (which sort of defeated the point of having the
wrapper class).