Properly handle defaults of CLI parameters
This commit improves naming consistency between command-line parameters,
config file names, and the parameters which GlmParser takes, so that the
amount of necessary code is reduced, and so that defaults can be handled
more cleanly.

This means that all config files have been modified, and string
interpolation in them is gone. We can achieve the same effect in
glm_parser.py if we need it, or using --params to override the config
file.

Debugging functionality has been temporarily removed, but will be
brought back in a follow-up commit.