Bug fix, spark train is working now, finally
The bug which occurs in spark_train mode was caused by an earlier
untested bug fix on weight vector, which was basically a type
conversion bug. The key of WeightVector should be of `str` type instead
of `tuple`, it would help a lot if the developer actually run some
tests on the modules they have modified, instead of having others to
clean up his own bloody mess.